refactor(rocm): name the ROCm python package rocm/, not _rocm/
flashinfer/_rocm/ was the only underscore-prefixed directory in the tree
-- the sole other underscore module, _version.py, is setuptools-scm
generated rather than a style choice. The leading underscore came from a
general Python convention for private modules, not from this repo's
conventions, and nothing depends on it: it does not prevent import, and
the public-API surface is filtered on non-underscore names either way.

rocm/ as a subfolder of the tree it specialises is the agreed convention
for ROCm-specific directories, already applied to include/flashinfer/rocm/.

Mechanical: git mv plus 18 reference updates across six flashinfer
modules and seven test files. flashinfer._rocm no longer resolves;
flashinfer.rocm does.
